requirements.txt
flask
opencv-python
face-recognition
numpy
ultralytics
mysql-connector-python
cryptography